feat: adding support for antlr in js
This is in preparation to migrate from python to javascript, with the
aim to make grammarizer a client side application with no need for a
server beyond serving static files.

In this commit we have added the antlr js runtime with help from npm
and have generated the js Visitor to use in the application, but it
remains unused for now.