feat: Create new webpage for project, whit syntax HL
In this commit a new simpler webpage is implemented. For this, new assets are
added to the project to supply fonts, styles and the syntax highlight
behaviour. This feature is provided by the Prism.js project, but the provided
css has been tweaked to support the implementation used by the page.

The implementation for the page is the overlay of a text area with invisible
text over a preformated code area that is kept in sync with the contents of
the text area by js. This is done because the textarea is the appropiate
element for text editing interfaces, but its text cant be formatted for SHL.