feat: Implements syntax highlight
The mechanism for syntax highlight consists of a textarea
overlayed over a pre+code element. This arrangement allows
interaction from the user with the convenient textarea while
showing the user the highlighted text from the pre element.

To allow this, the text area and it's contents are made transparent,
and the dimensions and positioning of the elements are made identical.
JS functions as event handlers allow keeping the text contents and
scroll statuses of the elements synchronized.

Syntax highlight styles and functionality provided by the Prism.js
library: prismjs.com/

Schema for editable highlighted text found at CSS-Tricks:
css-tricks.com/creating-an-editable-textarea-that-supports-syntax-highlighted-code/