feat: Adds support for ANTLR4 in the browser
Due to it's use of advanced ES6 proposed features (and possibly
other reasons as well) the ANTLR4 runtime for JS needs to be
bundled using webpack+babel.

Therefore this commit adds the proper config files for these tools
npm-installs the required dependecies, updates the .gitignore to
avoid node_modules and creates the appropiate entry point to
verify the bundling process.

In order to do this, the ANTLR cli tool is used to create the
lexer and parser for the EBNF grammar.