feat: Implements the translation algorithm
The functionality offered by this application is the translation
of EBNF grammars to BNF ones. This commit implements the algorithm
for doing so with the assistance of ANTLR4.

This algorithm consists of building a series of trees of nodes
representing the EBNF productions, with their respective parts.
Then reducing over these trees in accordance of the relations
between the nodes (a node's children can be mutually exclusive
alternatives, or concatenated sub rules) to produce the BNF
productions as strings.

Finally, the algorithm has been bundled by webpack, exposing it
as the transform function in the EBNF module.